refactor: import OID constants from the _oids module
