+++ psycopg/psycopg/types/array.py
@@ -13,10 +13,10 @@ from .. import errors as e
 from .. import postgres
 from ..abc import AdaptContext, Buffer, Dumper, DumperKey, NoneType, Loader, Transformer
 from ..adapt import RecursiveDumper, RecursiveLoader, PyFormat
+from .._oids import TEXT_OID, INVALID_OID
 from .._compat import cache, prod
 from .._struct import pack_len, unpack_len
 from .._cmodule import _psycopg
-from ..postgres import TEXT_OID, INVALID_OID
 from .._typeinfo import TypeInfo
 
 _struct_head = struct.Struct("!III")  # ndims, hasnull, elem oid
