+++ psycopg/psycopg/types/multirange.py
@@ -14,8 +14,8 @@ from .. import postgres
 from ..pq import Format
 from ..abc import AdaptContext, Buffer, Dumper, DumperKey
 from ..adapt import RecursiveDumper, RecursiveLoader, PyFormat
+from .._oids import INVALID_OID, TEXT_OID
 from .._struct import pack_len, unpack_len
-from ..postgres import INVALID_OID, TEXT_OID
 from .._typeinfo import MultirangeInfo as MultirangeInfo  # exported here
 
 from .range import Range, T, load_range_text, load_range_binary
